rename controlle manager and route
