+++ src/View/Category/index.html.twig
@@ -1,0 +1,39 @@
+{% extends 'layout.html.twig' %}
+
+{% block content %}
+
+	<div class="row">
+		<h1 class="col-12 text-center mx-auto my-5 " style="width: fit-content; ">
+			Nous y sommes
+			{% for user in users|slice(0,1) %}
+				{{user.nickname}}
+			{% endfor %}
+			,
+																																																																																																				choisis un thème pour lancer la partie !</h1>
+
+	</div>
+	<div class="container">
+
+		<div class="row justify-content-center">
+			{% for categorie in categories %}
+
+				<div class="  col-sm-12 col-lg-6">
+
+					<a href="/game/category?id={{categorie.id}}">
+						<div class="card bg-dark shadow text-white m-3">
+							<img src="/assets/images/{{categorie.image}}" class="card-img img-category" alt="...">
+							<div class="card-img-overlay">
+								<h1 class="card-title">{{categorie.name}}
+								</h1>
+								<p class="card-text d-none d-md-block"></p>
+							</div>
+						</div>
+					</a>
+				</div>
+
+
+			{% endfor %}
+
+		</div>
+	</div>
+{% endblock %}
