--- src/View/Game/index.html.twig
@@ -1,25 +1,0 @@
-{% extends 'layout.html.twig' %}
-
-{% block content %}
-
-	<h1 class="mx-auto my-1 game-h1">
-
-		Catégorie
-		{{categories.name}}
-
-	</h1>
-	<div class="answer-center">
-		<h2 class="col-12 mx-auto my-1" style='width:fit-content;'>Devine le titre du morceau :</h2>
-		{{track.name}}
-		<form action="" method="POST">
-			<label for="proposal"></label>
-			<input id="proposal" name="proposal" type="text" placeholder="votre réponse">
-			<div class="button-flex">
-				<input class="button-valid" type="submit" value="Valider">
-				<a class="pass" href="/game/track&id={{ track.id }}">Passer</a>
-			</div>
-		</form>
-
-
-	</div>
-{% endblock %}
