+++ src/View/Quizz/index.html.twig
@@ -2,38 +2,24 @@
 
 {% block content %}
 
-	<div class="row">
-		<h1 class="col-12 text-center mx-auto my-5 " style="width: fit-content; ">
-			Nous y sommes
-			{% for user in users|slice(0,1) %}
-				{{user.nickname}}
-			{% endfor %}
-			,
-																																																																																																				choisis un thème pour lancer la partie !</h1>
+	<h1 class="mx-auto my-1 game-h1">
 
-	</div>
-	<div class="container">
+		Catégorie
+		{{categories.name}}
 
-		<div class="row justify-content-center">
-			{% for categorie in categories %}
-
-				<div class="  col-sm-12 col-lg-6">
-
-					<a href="/game/category?id={{categorie.id}}">
-						<div class="card bg-dark shadow text-white m-3">
-							<img src="/assets/images/{{categorie.image}}" class="card-img img-category" alt="...">
-							<div class="card-img-overlay">
-								<h1 class="card-title">{{categorie.name}}
-								</h1>
-								<p class="card-text d-none d-md-block"></p>
-							</div>
-						</div>
-					</a>
-				</div>
+	</h1>
+	<div class="answer-center">
+		<h2 class="col-12 mx-auto my-1" style='width:fit-content;'>Devine le titre du morceau :</h2>
+		{{track.name}}
+		<form action="" method="POST">
+			<label for="proposal"></label>
+			<input id="proposal" name="proposal" type="text" placeholder="votre réponse">
+			<div class="button-flex">
+				<input class="button-valid" type="submit" value="Valider">
+				<a class="pass" href="/game/track&id={{ track.id }}">Passer</a>
+			</div>
+		</form>
 
 
-			{% endfor %}
-
-		</div>
 	</div>
 {% endblock %}
